pyproject.toml
[tool.poetry]
name = "mercari-api"
version = "0.1.0"
description = ""
license = "MIT"
authors = []

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
boto3 = "^1.21.24"
django-storages = "^1.12.3"
django-cors-headers = "^3.11.0"
sentry-sdk = "^1.5.8"
redis = "^4.2.0"
python-decouple = "^3.6"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
